chore: Remove typescript from mock package for simplicity
There were various issues with bundling this, so it is easier to just manually provide the types.
