+++ package.json
@@ -5,8 +5,7 @@
   "repository": "https://github.com/fabscale/ember-graphql-client",
   "workspaces": [
     "packages/client",
-    "packages/mock",
-    "packages/release-it-generate-types"
+    "packages/mock"
   ],
   "scripts": {
     "lint:js": "eslint . --cache"
@@ -14,7 +13,6 @@
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "~4.28.0",
     "@typescript-eslint/parser": "~4.28.0",
-    "@ember-graphql-client/release-it-generate-types": "*",
     "babel-eslint": "~10.1.0",
     "eslint": "~7.31.0",
     "eslint-config-prettier": "~8.3.0",
@@ -33,7 +31,6 @@
   },
   "release-it": {
     "plugins": {
-      "@ember-graphql-client/release-it-generate-types": {},
       "release-it-lerna-changelog": {
         "infile": "CHANGELOG.md"
       },
