+++ packages/client/package.json
@@ -26,8 +26,8 @@
     "test:ember": "ember test",
     "test:ember-one": "ember try:one",
     "test:ember-compatibility": "ember try:each",
-    "ts:precompile": "ember ts:precompile",
-    "ts:clean": "ember ts:clean"
+    "prepack": "ember ts:precompile",
+    "postpack": "ember ts:clean"
   },
   "dependencies": {
     "@ember/test-waiters": "^2.4.5",
