--- packages/mock/addon/mock-graphql-error.ts
@@ -1,30 +1,0 @@
-import { GraphQLError } from 'graphql';
-import type { Source } from 'graphql';
-
-interface GraphQLErrorInterface {
-  message: string;
-  source?: Source;
-  positions?: ReadonlyArray<number>;
-  path?: ReadonlyArray<string | number>;
-  originalError?: Error;
-  extensions?: { [key: string]: any };
-}
-
-export function mockGraphQLError({
-  message,
-  source,
-  positions,
-  path,
-  originalError,
-  extensions,
-}: GraphQLErrorInterface): GraphQLError {
-  return new GraphQLError(
-    message,
-    undefined,
-    source,
-    positions,
-    path,
-    originalError,
-    extensions
-  );
-}
