+++ packages/mock/package.json
@@ -18,16 +18,13 @@
     "lint:fix": "npm-run-all --aggregate-output --continue-on-error --parallel lint:*:fix",
     "lint:hbs": "ember-template-lint .",
     "lint:hbs:fix": "ember-template-lint . --fix",
-    "lint:ts": "tsc",
     "lint:js": "eslint . --cache",
     "lint:js:fix": "eslint . --fix",
     "start": "ember serve",
     "test": "npm-run-all lint test:*",
     "test:ember": "ember test",
     "test:ember-one": "ember try:one",
-    "test:ember-compatibility": "ember try:each",
-    "ts:precompile": "ember ts:precompile",
-    "ts:clean": "ember ts:clean"
+    "test:ember-compatibility": "ember try:each"
   },
   "dependencies": {
     "@ember-graphql-client/client": "0.1.5",
@@ -35,7 +32,6 @@
     "@graphql-tools/utils": "^7.10.0",
     "ember-auto-import": "^2.0.2",
     "ember-cli-babel": "^7.26.6",
-    "ember-cli-typescript": "^4.2.1",
     "graphql": "~15.5.1",
     "graphql-request": "^3.4.0"
   },
@@ -70,7 +66,6 @@
     "ember-cli-inject-live-reload": "~2.1.0",
     "ember-cli-sri": "~2.1.1",
     "ember-cli-terser": "~4.0.2",
-    "ember-cli-typescript-blueprints": "~3.0.0",
     "ember-disable-prototype-extensions": "~1.1.3",
     "ember-export-application-global": "~2.0.1",
     "ember-load-initializers": "~2.1.2",
@@ -84,7 +79,6 @@
     "npm-run-all": "~4.1.5",
     "qunit": "~2.16.0",
     "qunit-dom": "~1.6.0",
-    "typescript": "~4.3.5",
     "webpack": "~5.45.1"
   },
   "engines": {
