+++ packages/mock/tests/dummy/app/components/posts.js
@@ -1,24 +1,21 @@
 import { inject as service } from '@ember/service';
 import Component from '@glimmer/component';
 import { tracked } from '@glimmer/tracking';
-import GraphQLService from '@ember-graphql-client/client/services/graphql';
 import query from 'dummy/gql/queries/posts.graphql';
 
-type Args = any;
+export default class PostComponent extends Component {
+  @service graphql;
 
-export default class PostComponent extends Component<Args> {
-  @service declare graphql: GraphQLService;
-
-  @tracked posts?: any;
+  @tracked posts;
   @tracked page = 1;
 
-  constructor(owner: unknown, args: Args) {
+  constructor(owner, args) {
     super(owner, args);
 
     this._loadPosts();
   }
 
-  async _loadPosts(): Promise<void> {
+  async _loadPosts() {
     let postsData = await this.graphql.query({
       query,
       variables: { pageSize: 10, page: this.page },
