--- packages/mock/tests/dummy/app/config/environment.d.ts
@@ -1,14 +1,0 @@
-export default config;
-
-/**
- * Type declarations for
- *    import config from 'my-app/config/environment'
- */
-declare const config: {
-  environment: string;
-  modulePrefix: string;
-  podModulePrefix: string;
-  locationType: string;
-  rootURL: string;
-  APP: Record<string, unknown>;
-};
