+++ packages/mock/tests/dummy/app/post/route.js
@@ -1,7 +1,7 @@
 import Route from '@ember/routing/route';
 
 export default class PostRoute extends Route {
-  model(params: { id: string }): string | number {
+  model(params) {
     return params.id;
   }
 }
