+++ packages/mock/tests/test-helper.js
@@ -1,4 +1,3 @@
-// @ts-ignore
 import Application from 'dummy/app';
 import config from 'dummy/config/environment';
 import * as QUnit from 'qunit';
