--- packages/release-it-generate-types/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "@ember-graphql-client/release-it-generate-types",
-  "version": "0.1.5",
-  "description": "Generate types for the mock package before publishing.",
-  "main": "index.js",
-  "private": true,
-  "peerDpendencies": {
-    "release-it": "~14.10.0"
-  },
-  "devDependencies": {
-    "release-it": "~14.10.0"
-  }
-}
